v0.1.5, added transformation function, refactoring
transformation is an experimental function to read a custom user function and output the results. the exact usage the user can refer to the example in the docstring. the original intention is to create a functional style (lambda only), but it's impractical to work with large scale data processing this way. however, it is possible that declarative style may need to unwanted side effect (read write from disk etc).